+++ account-data-api/src/se/liu/ida/tdp024/account/data/api/facade/AccountEntityFacade.java
@@ -1,8 +1,6 @@
 package se.liu.ida.tdp024.account.data.api.facade;
 
-import java.util.List;
 import se.liu.ida.tdp024.account.data.api.entity.Account;
-import se.liu.ida.tdp024.account.data.api.entity.Transaction;
 
 public interface AccountEntityFacade {
     
@@ -10,7 +8,7 @@ public interface AccountEntityFacade {
     
     public Account find(String name);
     
-    public String debit(int id,int debiteras);
+    public String debit(long id,long debiteras);
     
-    public String kredit(int id,int krediteras);
+    public String kredit(long id,long krediteras);
 }
